refactor(db): abstract LISH flag and transfer-stat persistence into own module
