+++ backend/src/db/lishs.ts
@@ -582,53 +582,5 @@ function getHaveChunksList(db: Database, internalID: number): string[] {
 		.map(r => r.checksum);
 }
 
-// -- Upload enabled persistence --
-
-export function setUploadEnabled(db: Database, lishID: LISHid, enabled: boolean): void {
-	db.run('UPDATE lishs SET upload_enabled = ? WHERE lish_id = ?', [enabled ? 1 : 0, lishID]);
-}
-
-export function setDownloadEnabled(db: Database, lishID: LISHid, enabled: boolean): void {
-	db.run('UPDATE lishs SET download_enabled = ? WHERE lish_id = ?', [enabled ? 1 : 0, lishID]);
-}
-
-export function getUploadEnabledLishs(db: Database): Set<string> {
-	return new Set(
-		db
-			.query<{ lish_id: string }, []>('SELECT lish_id FROM lishs WHERE upload_enabled = TRUE')
-			.all()
-			.map(r => r.lish_id)
-	);
-}
-
-export function getDownloadEnabledLishs(db: Database): Set<string> {
-	return new Set(
-		db
-			.query<{ lish_id: string }, []>('SELECT lish_id FROM lishs WHERE download_enabled = TRUE')
-			.all()
-			.map(r => r.lish_id)
-	);
-}
-
-// -- Transfer stats persistence --
-
-export function incrementUploadedBytes(db: Database, lishID: LISHid, bytes: number): void {
-	db.run('UPDATE lishs SET total_uploaded_bytes = total_uploaded_bytes + ? WHERE lish_id = ?', [bytes, lishID]);
-}
-
-export function incrementDownloadedBytes(db: Database, lishID: LISHid, bytes: number): void {
-	db.run('UPDATE lishs SET total_downloaded_bytes = total_downloaded_bytes + ? WHERE lish_id = ?', [bytes, lishID]);
-}
-
-export function getTransferStats(db: Database, lishID: LISHid): { uploadedBytes: number; downloadedBytes: number } {
-	const row = db.query<{ total_uploaded_bytes: number; total_downloaded_bytes: number }, [string]>('SELECT total_uploaded_bytes, total_downloaded_bytes FROM lishs WHERE lish_id = ?').get(lishID);
-	return { uploadedBytes: row?.total_uploaded_bytes ?? 0, downloadedBytes: row?.total_downloaded_bytes ?? 0 };
-}
-
-export function setLISHError(db: Database, lishID: LISHid, errorCode: string, errorDetail?: string): void {
-	db.run('UPDATE lishs SET error_code = ?, error_detail = ? WHERE lish_id = ?', [errorCode, errorDetail ?? null, lishID]);
-}
-
-export function clearLISHError(db: Database, lishID: LISHid): void {
-	db.run('UPDATE lishs SET error_code = NULL, error_detail = NULL WHERE lish_id = ?', [lishID]);
-}
+// Enabled-flag, transfer-stat, and error-state persistence live in lishs-flags.ts; re-exported via this barrel.
+export * from './lishs-flags.ts';
